requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
camel-ai==0.2.44
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
docstring-parser==0.15
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mcp==1.6.0
numpy==2.2.4
openai==1.73.0
pandas==2.2.3
pillow==11.1.0
psutil==5.9.8
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
-e git+https://github.com/ris3abh/ResumeForgeAI.git@90abeaff1160d8ede385ed39d4e2dd81e9592542#egg=resume_optimizer
rpds-py==0.24.0
six==1.17.0
sniffio==1.3.1
sse-starlette==2.2.1
starlette==0.46.1
tiktoken==0.7.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.0
